a function that executes a function given as a parameter on each element of an array

+++ function_pointers/1-array_iterator.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - fonction pointer that uses fonction on each element of the array
+ * array_iterator - uses fonction on each element of the array
  * @array: the array
  * @size: size of the array
  * @action: the fonction pointer
@@ -9,8 +9,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-    int i;
+	unsigned long int i;
 
-    for (i = 0 ; i < size ; i++)
-        action(array[i]);
+	for (i = 0 ; i < size ; i++)
+		action(array[i]);
 }
